Migrate tests to more appropriate class
